+++ app/Http/Controllers/Api/ApplicantEmploymentController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers\Api;
 
+use App\Module;
 use App\Applicant;
+use App\ApplicantEducation;
 use App\ApplicantEmployment;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
@@ -11,13 +13,31 @@ use App\Http\Requests\ApplicantEmploymentRequest;
 
 class ApplicantEmploymentController extends Controller
 {
+    protected $module;
+
+    public function __construct()
+    {
+        $this->module = Module::where('code', 'applicant')->first();
+    }
+
     public function index(Applicant $applicant)
     {
-        return $applicant->employments;
+        $this->authorize('allows', [$this->module, 'view']);
+
+        return $applicant->employments()->orderBy('id', 'desc')->get();
+    }
+
+    public function show(Applicant $applicant, ApplicantEmployment $employment)
+    {
+        $this->authorize('allows', [$this->module, 'show']);
+
+        return $employment;
     }
 
     public function store(ApplicantEmploymentRequest $request, Applicant $applicant, ApplicantEmployment $employment)
     {
+        $this->authorize('allows', [$this->module, 'create']);
+
         $employment = $applicant->employments()->create($request->only($employment->fillable));
 
         Log::info(auth()->user()->username . ' - Applicant Employment Created', [
@@ -32,6 +52,8 @@ class ApplicantEmploymentController extends Controller
 
     public function update(ApplicantEmploymentRequest $request, Applicant $applicant, ApplicantEmployment $employment)
     {
+        $this->authorize('allows', [$this->module, 'update']);
+
         $employment->update($request->only($employment->fillable));
 
         Log::info(auth()->user()->username . ' - Applicant Employment Updated', [
@@ -46,6 +68,8 @@ class ApplicantEmploymentController extends Controller
 
     public function destroy(Applicant $applicant, ApplicantEmployment $employment)
     {
+        $this->authorize('allows', [$this->module, 'delete']);
+
         $employment->delete();
 
         Log::info(auth()->user()->username . ' - Applicant Employment Deleted', [
